% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrtpoints.R
\name{gdal_grid}
\alias{gdal_grid}
\title{GDAL grid}
\usage{
gdal_grid(
  pts,
  dimension = c(256, 256),
  extent = NULL,
  algorithm = "linear",
  read = TRUE,
  options = NULL
)
}
\arguments{
\item{pts}{points, columns of X, Y, Z}

\item{dimension}{ncols, nrows}

\item{extent}{xmin,xmax,ymin,ymax}

\item{algorithm}{one of invdist, invdistnn, average, linear see Details}
}
\value{
matrix of raster values, numeric
}
\description{
Runs the gdal_grid utility.
}
\details{
Algorithms listed at https://gdal.org/programs/gdal_grid.html#interpolation-algorithmsuse

Pass in paramaters the algorithm name, e. g. 'invdist:power=2.0:smoothing=1.0' - see examples.
}
\examples{
n <- 500
rc <- cbind(sample(nrow(volcano), n, replace = TRUE), sample(ncol(volcano), n, replace = TRUE))
xyz <- cbind(rc, volcano[rc])
ex <- c(range(xyz[,1]), range(xyz[,2]))
v <- gdal_grid(xyz, extent = ex)

ximage::ximage(v, extent = ex, asp = 1)

v1 <- gdal_grid(xyz, extent = ex, algorithm = "invdist:power=2.0:smoothing=1.0 ")
}
