% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrtpoints.R
\name{vrtpoints}
\alias{vrtpoints}
\title{Write points xyz to CSV for GDAL.}
\usage{
vrtpoints(pts, name = NULL)
}
\arguments{
\item{pts}{data frame or matrix of points, X, Y, Z}

\item{name}{optional, used as the layername in the GDAL source}
}
\value{
path to tempfile
}
\description{
Obtain a filename that wraps a CSV file with the input points in it. This is suitable
for use with \url{gdal_grid} the command line utility, but see \code{\link[=gdal_grid]{gdal_grid()}} the function
for a convenient wrapper.
}
\examples{
n <- 500
rc <- cbind(sample(nrow(volcano), n, replace = TRUE), sample(ncol(volcano), n, replace = TRUE))
xyz <- cbind(rc, volcano[rc])
plot(xyz[,2:1], col = palr::d_pal(xyz[,3]), pch = 19, cex = .3)
file <- vrtpoints(xyz)
sf::gdal_utils("grid", file, tf <- tempfile(fileext = ".tif"))
sf::read_sf(file)
sf::gdal_utils("info", tf)
}
